+++ packages/nimble-components/src/table-column/enum-base/models/mapping-config.ts
@@ -2,5 +2,5 @@
  * Common state shared across Mapping Config
  */
 export abstract class MappingConfig {
-    public constructor(public readonly text: string) {}
+    public constructor(public readonly text: string | undefined) {}
 }