+++ packages/nimble-components/src/table-column/text-base/cell-view/template.ts
@@ -2,12 +2,19 @@ import { html } from '@microsoft/fast-element';
 
 import type { TableColumnTextCellViewBase } from '.';
 import { overflow } from '../../../utilities/directive/overflow';
+import { TextCellViewBaseAlignment } from './types';
 
 export const template = html<TableColumnTextCellViewBase>`
-    <span
-        ${overflow('hasOverflow')}
-        title=${x => (x.hasOverflow && x.text ? x.text : null)}
+    <template
+        class="${x => (x.alignment === TextCellViewBaseAlignment.right
+        ? 'right-align'
+        : '')}"
     >
-        ${x => x.text}
-    </span>
+        <span
+            ${overflow('hasOverflow')}
+            title=${x => (x.hasOverflow && x.text ? x.text : null)}
+        >
+            ${x => x.text}
+        </span>
+    </template>
 `;